requirements.txt
# Agent Framework
crewai>=0.28.0
langgraph>=0.1.0
langchain>=0.2.0
langchain-community>=0.2.0

# LLM
groq>=0.4.0
ollama>=0.1.0

# RAG / Vector Store
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Airflow Client
apache-airflow-client>=2.7.0

# API / Web
fastapi>=0.110.0
uvicorn>=0.27.0
requests>=2.31.0

# Notifications
slack-sdk>=3.27.0

# Utils
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi


uvicorn[standard]
crewai
crewai-tools
langgraph
langchain-groq
langchain-core
